[PropertyAccess] Extract AccessInfoGuesser from PropertyAccessor
